kernel: Use tiles and move to local memory
I now use 12 tiles to compute the sobel, each tile is 1600 pixels long.
The pixels needed for each tile are brought into the local memory and the
computation is done with the local versions. The outputs are stored in
local memory also, but the memory is for all tiles. So the output memory
is local to the kernel and global to the tiles.